refactor: rename auth-form to AuthForm
