+++ client/components/index.js
@@ -8,7 +8,7 @@ export {default as MapAllItems} from './MapAllItems'
 export {default as MapSingleItem} from './MapSingleItem'
 export {default as Navbar} from './Navbar'
 export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {Login, Signup} from './AuthForm'
 export {default as Items} from './Items'
 export {default as temp} from './temp'
 export {default as NewItemForm} from './NewItemForm'
